requirements.txt
Pillow
vidspinner
